# src/auth/controller/responses.go
package controller

import "github.com/CPU-commits/Template_Go-EventDriven/src/auth/model"

type LoginResponse struct {
	AccessToken  string      `json:"accessToken" example:"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"`
	RefreshToken string      `json:"refreshToken" example:"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"`
	User         *model.User `json:"user"`
}

type RecoveryCodeResponse struct {
	Recovery *model.Recovery `json:"recovery"`
}

type VerifyRecoveryCodeResponse struct {
	Exists bool   `json:"exists"`
	Token  string `json:"token"`
}